scripts/requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
